% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{assign_y_levels_genes}
\alias{assign_y_levels_genes}
\title{determine vertical coordinates of genes in gene model plot}
\usage{
assign_y_levels_genes(df)
}
\arguments{
\item{df}{dataframe of gene chromosome, start coordinate, and end coordinate}
}
\value{
dataframe complete with y coordinates for each gene in gene model plot
}
\description{
determine vertical coordinates of genes in gene model plot
}
