% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{assign_y_levels_transcripts}
\alias{assign_y_levels_transcripts}
\title{determine vertical coordinates of transcripts in gene model plot}
\usage{
assign_y_levels_transcripts(df)
}
\arguments{
\item{df}{dataframe of gene and transcript chromosome, start coordinate, and end coordinate}
}
\value{
dataframe complete with y coordinates for each gene and transcript in gene model plot
}
\description{
determine vertical coordinates of transcripts in gene model plot
}
