% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{get_chrom_names}
\alias{get_chrom_names}
\title{obtain list of chromosomes in a bigwig, bed, bedgraph, txt or tsv file}
\usage{
get_chrom_names(file)
}
\arguments{
\item{file}{filepath for bigwig, bed, bedgraph, txt or tsv file}
}
\value{
chromosomes in file
}
\description{
obtain list of chromosomes in a bigwig, bed, bedgraph, txt or tsv file
}
