% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{get_plot_title}
\alias{get_plot_title}
\title{obtain plot title from generated plot}
\usage{
get_plot_title(p)
}
\arguments{
\item{p}{ggplot object}
}
\value{
plot title
}
\description{
obtain plot title from generated plot
}
