% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{plot_coverage}
\alias{plot_coverage}
\title{plot coverage tracks}
\usage{
plot_coverage(
  genomicLoc,
  covFiles,
  covTrackNames = NULL,
  covTrackColors = "black",
  ymin = NULL,
  ymax = NULL,
  fillArea = TRUE,
  logScale = FALSE,
  rasterizePlot = FALSE,
  fontSize = 9
)
}
\arguments{
\item{genomicLoc}{character vector of the form c(chromosome, start coordinate, end coordinate) (ex: c("chrX", 24040226, 24232664))}

\item{covFiles}{character vector containing a list of file paths to any coverage files (typically in bigWig or bedGraph format)}

\item{covTrackNames}{character vector containing the label of each coverage track. Should be the same length as covFiles and in the same order. By default it will label them as "Cov_1", "Cov_2", etc}

\item{covTrackColors}{character vector containing the color of each coverage track. If only one color is provided, all tracks will be that color. If different tracks must be different colors, specify the color for each track in order. Default is "black"}

\item{ymin}{numeric, minimum value for y-axis for coverage plot}

\item{ymax}{numeric, maximum value for y-axis for coverage plot}

\item{fillArea}{boolean, set this to FALSE if you do not want to fill in the area under your coverage tracks. Default is TRUE}

\item{logScale}{boolean, set this to TRUE if you want data plotted on a log scale. Default is FALSE}

\item{rasterizePlot}{boolean, set this to TRUE if you want to rasterize the coverage plot. This can be useful as coverage files can be rather large, and that can mess up the plot when you save it as a vector file. Default is FALSE}

\item{fontSize}{a numeric for desired font size. Default is 9.}
}
\value{
ggplot of coverage tracks
}
\description{
plot coverage tracks
}
