% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{plot_gene}
\alias{plot_gene}
\title{plot gene models}
\usage{
plot_gene(
  genomicLoc,
  includeTranscripts = FALSE,
  includeTxtNames = TRUE,
  transcript_list = NULL,
  supportedTranscriptsOnly = TRUE,
  transcript_filters = c("transcript_gencode_basic"),
  appris_options = NULL,
  tsl_options = NULL,
  tag_options = NULL,
  canonicalTranscriptOnly = FALSE,
  mart = NULL,
  ensembl_set = "hsapiens_gene_ensembl",
  gene_symbol = "hgnc_symbol",
  includeNPC = FALSE,
  custom_anno = NULL,
  upDown = c(2000, 2000)
)
}
\arguments{
\item{genomicLoc}{character vector. A gene name (ex:"ZFX") OR a genomic locus of the form c(chromosome, start coordinate, end coordinate) (ex: c("chrX", 24040226, 24232664))}

\item{includeTranscripts}{boolean. Whether to plot transcripts. Default is FALSE.}

\item{includeTxtNames}{boolean. Whether to label transcript models. If set to TRUE, requires includeGenome and include Transcripts to be set to TRUE. Default is FALSE.}

\item{transcript_list}{list, used to specify the names of the transcripts you want to include (if you only want a specific set of transcripts). Requires includeTranscripts to be set to TRUE. overrides any specified transcript filters}

\item{supportedTranscriptsOnly}{boolean, whether to only include well-supported transcripts as defined by the 'transcript_filters' parameter. Default is TRUE.}

\item{transcript_filters}{character vector of which transcript filters to clear when supportedTranscriptsOnly is set to TRUE. Common choices from ensembl are 'transcript_gencode_basic', 'transcript_appris', and 'transcript_tsl'. the 'appris_options' and 'tsl_options' parameters allow you to specify which APPRIS classification or transcript support levels, respectively, that you want to keep. Any other filter will simply keep transcripts that have a non-empty field for that filter. Default is 'transcript_gencode_basic'}

\item{appris_options}{character vector of APPRIS filters to implement if 'transcript_appris' is provided in 'transcript_filters'. can specify both principal (P1-P5) and alternative transcripts (A1-A5). ex: c("P1","P2","A1") will give you principal 1, principal2, and alternative 1 transcripts. If 'transcript_appris' is provided as 'transcript_filter' and 'appris_options' is not specified, all principal transcripts will be kept}

\item{tsl_options}{character vector of transcript support levels you want to keep if 'transcript_tsl' is provided in 'transcript_filters'}

\item{tag_options}{character vector of which tags to keep if using a custom annotation and setting supportedTranscriptsOnly to TRUE.}

\item{canonicalTranscriptOnly}{boolean, whether to only show the canoncial transcript. Default is FALSE.}

\item{mart}{a mart obtained by utilizing the useMart function in biomaRt}

\item{ensembl_set}{if 'mart' is not provided, which Ensembl annotation you want to pull from. Default is 'hsapiens_gene_ensembl'. Most ensembl sets follow this format; for example if you're using mouse it would be 'mmusculus_gene_ensembl'}

\item{gene_symbol}{nomenclature system you are using for genes. Default is 'hgnc_symbol'but switches to 'external_gene_name' for any ensembl_set that is not 'hsapiens_gene_ensembl'. If you are using other nomenclature systems (ex: 'mgi_symbol' for mouse), you must specify. If using a custom annotation file, you likely need to use 'gene_id'.}

\item{includeNPC}{boolean, whether to include genomic features that are not protein coding. Default is FALSE.}

\item{custom_anno}{file path to a .gtf/.gff if using a custom annotation rather than ensembl}

\item{upDown}{a character vector of length 2. if genomicLoc is a gene, how many base pairs upstream (first value) and downstream (second value) of the gene you want in your figure. Default is c(2000,2000)}
}
\value{
list of length 2: figure (ggplot of gene models); genePlot (ggplot of gene models)
}
\description{
plot gene models
}
