% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{plot_genomic_tracks}
\alias{plot_genomic_tracks}
\title{plot gene models, coverage tracks, peaks, and/or loops across region of interest}
\usage{
plot_genomic_tracks(
  genomicLoc,
  includeGenome = TRUE,
  includeTranscripts = FALSE,
  includeTxtNames = TRUE,
  transcript_list = NULL,
  supportedTranscriptsOnly = TRUE,
  transcript_filters = c("transcript_gencode_basic"),
  appris_options = NULL,
  tsl_options = NULL,
  tag_options = NULL,
  canonicalTranscriptOnly = FALSE,
  mart = NULL,
  ensembl_set = "hsapiens_gene_ensembl",
  gene_symbol = "hgnc_symbol",
  includeNPC = FALSE,
  custom_anno = NULL,
  upDown = c(2000, 2000),
  genome_position = "bottom",
  covFiles = NULL,
  covTrackNames = NULL,
  covTrackColors = "black",
  ymin = NULL,
  ymax = NULL,
  fillArea = TRUE,
  logScale = FALSE,
  rasterizeCovPlot = FALSE,
  peakFiles = NULL,
  peakTrackNames = NULL,
  labelAllPeaks = FALSE,
  peakTrackColors = "black",
  specialPeaks = NULL,
  labelSpecialPeaks = FALSE,
  specialPeakColors = NULL,
  labelStrand = FALSE,
  strandColors = NULL,
  loopFiles = NULL,
  loopTrackNames = NULL,
  loopTrackColors = "black",
  lineSize = 0.8,
  alpha = 0.8,
  minScore = NULL,
  rasterizeLoopPlot = FALSE,
  specialLoops = NULL,
  specialLoopColors = NULL,
  loop_orientation = "above",
  trackOrder_type = c("coverage", "peaks", "loops", "genome"),
  fontSize = 9,
  saveFigure = FALSE,
  figureName = "plot_genomic_tracks_figure",
  figureFormat = "png"
)
}
\arguments{
\item{genomicLoc}{character vector. A gene name (ex:"ZFX") OR a genomic locus of the form c(chromosome, start coordinate, end coordinate) (ex: c("chrX", 24040226, 24232664))}

\item{includeGenome}{boolean. Whether to include gene models in the final figure. Almost always set to TRUE. Default is TRUE.}

\item{includeTranscripts}{boolean. Whether to plot transcripts. Default is FALSE.}

\item{includeTxtNames}{boolean. Whether to label transcript models. If set to TRUE, requires includeGenome and include Transcripts to be set to TRUE. Default is FALSE.}

\item{transcript_list}{list, used to specify the names of the transcripts you want to include (if you only want a specific set of transcripts). Requires includeTranscripts to be set to TRUE. overrides any specified transcript filters}

\item{supportedTranscriptsOnly}{boolean, whether to only include well-supported transcripts as defined by the 'transcript_filters' parameter. Default is TRUE.}

\item{transcript_filters}{character vector of which transcript filters to clear when supportedTranscriptsOnly is set to TRUE. Common choices from ensembl are 'transcript_gencode_basic', 'transcript_appris', and 'transcript_tsl'. the 'appris_options' and 'tsl_options' parameters allow you to specify which APPRIS classification or transcript support levels, respectively, that you want to keep. Any other filter will simply keep transcripts that have a non-empty field for that filter. Default is 'transcript_gencode_basic'}

\item{appris_options}{character vector of APPRIS filters to implement if 'transcript_appris' is provided in 'transcript_filters'. can specify both principal (P1-P5) and alternative transcripts (A1-A5). ex: c("P1","P2","A1") will give you principal 1, principal2, and alternative 1 transcripts. If 'transcript_appris' is provided as 'transcript_filter' and 'appris_options' is not specified, all principal transcripts will be kept}

\item{tsl_options}{character vector of transcript support levels you want to keep if 'transcript_tsl' is provided in 'transcript_filters'}

\item{tag_options}{character vector of which tags to keep if using a custom annotation and setting supportedTranscriptsOnly to TRUE.}

\item{canonicalTranscriptOnly}{boolean, whether to only show the canoncial transcript. Default is FALSE.}

\item{mart}{a mart obtained by utilizing the useMart function in biomaRt}

\item{ensembl_set}{if 'mart' is not provided, which Ensembl annotation you want to pull from. Default is 'hsapiens_gene_ensembl'. Most ensembl sets follow this format; for example if you're using mouse it would be 'mmusculus_gene_ensembl'}

\item{gene_symbol}{nomenclature system you are using for genes. Default is 'hgnc_symbol'but switches to 'external_gene_name' for any ensembl_set that is not 'hsapiens_gene_ensembl'. If you are using other nomenclature systems (ex: 'mgi_symbol' for mouse), you must specify. If using a custom annotation file, you likely need to use 'gene_id'.}

\item{includeNPC}{boolean, whether to include genomic features that are not protein coding. Default is FALSE.}

\item{custom_anno}{file path to a .gtf/.gff if using a custom annotation rather than ensembl}

\item{upDown}{a character vector of length 2. if genomicLoc is a gene, how many base pairs upstream (first value) and downstream (second value) of the gene you want in your figure. Default is c(2000,2000)}

\item{genome_position}{if includeGenome is set to TRUE but 'genome' is not in 'trackOrder_type', where do you want gene models in your figure. set to "bottom" if you want the genome to be at the bottom of the figure; "top" if you want it to be at the top. Default is 'bottom'}

\item{covFiles}{character vector containing a list of file paths to any coverage files (typically in bigWig or bedGraph format)}

\item{covTrackNames}{character vector containing the label of each coverage track. Should be the same length as covFiles and in the same order. By default it will label them as "Cov_1", "Cov_2", etc}

\item{covTrackColors}{character vector containing the color of each coverage track. If only one color is provided, all tracks will be that color. If different tracks must be different colors, specify the color for each track in order. Default is "black"}

\item{ymin}{numeric, minimum value for y-axis for coverage plot}

\item{ymax}{numeric, maximum value for y-axis for coverage plot}

\item{fillArea}{boolean, set this to FALSE if you do not want to fill in the area under your coverage tracks. Default is TRUE}

\item{logScale}{boolean, set this to TRUE if you want data plotted on a log scale. Default is FALSE}

\item{rasterizeCovPlot}{boolean, set this to TRUE if you want to rasterize the coverage plot. This can be useful as coverage files can be rather large, and that can mess up the plot when you save it as a vector file. Default is FALSE}

\item{peakFiles}{character vector containing a list of file paths to any peak files (.bed format)}

\item{peakTrackNames}{character vector containing the label of each peak track. Should be the same length as 'peakFiles' and in the same order. By default it will label them as "Peaks_1", "Peaks_2", etc}

\item{labelAllPeaks}{boolean, set to TRUE if each individual peak should be named. assumes names are provided in the fourth column of a given .bed file}

\item{peakTrackColors}{character vector containing color of each peak track. If only one color is provided, all tracks will be that color. If different tracks must be different colors, specify the color for each track in order. Default is "black"}

\item{specialPeaks}{character vector listing the names of specific peaks you want to either label or put in a different color}

\item{labelSpecialPeaks}{boolean, set to TRUE if each special peak should be named. requires 'specialPeaks' to be specified}

\item{specialPeakColors}{character vector containing the color(s) you want to give your special peaks. requires 'specialPeaks' to be specified}

\item{labelStrand}{boolean, set to TRUE if the strand of each peak should be indicated. assumes strand information is in the sixth column of a given .bed file}

\item{strandColors}{character vector of length 2. First is the color for the + strand, the second is the color for the - strand. overrides 'peakTrackColors' but does NOT override 'specialPeakColors'}

\item{loopFiles}{character vector containing a list of file paths to any contact/loop files (.bedpe format)}

\item{loopTrackNames}{character vector containing the label of each loop track. Should be the same length as 'contactFiles' and in the same order. By default it will label them as "Loops_1", "Loops_2", etc}

\item{loopTrackColors}{character vector containing color of each loop track. If only one color is provided, all tracks will be that color. If different tracks must be different colors, specify the color for each track in order. Default is 'black'}

\item{lineSize}{numeric, width of lines used to draw loops. Default is 0.8}

\item{alpha}{numeric, alpha value (transparency) for loops. Default is 0.8}

\item{minScore}{numeric; any loops with scores lower than this will be thinner and more transparent, making the loops that pass this threshold more visible}

\item{rasterizeLoopPlot}{boolean, set this to TRUE if you want to rasterize the loop plot. This can be useful as loop files can be rather large, and that can mess up the plot when you save it as a vector file. Default is FALSE}

\item{specialLoops}{character vector listing the names of specific loops you want to put in a different color}

\item{specialLoopColors}{character vector containing the color(s) you want to give your special loops. requires 'specialLoops' to be specified}

\item{loop_orientation}{either "above" or "below". "above" will draw loops above a horizontal axis; "below" will draw loops below a horizontal axis. Default is "above'}

\item{trackOrder_type}{character vector with the desired order of track types. Default is c("coverage","peaks","loops", "genome")}

\item{fontSize}{a numeric for desired font size. Default is 9.}

\item{saveFigure}{boolean, whether you want to save the final figure to a file. Default is FALSE.}

\item{figureName}{if 'saveFigure' is TRUE, what you want the filename to be. Default is "plot_genomic_tracks_figure"}

\item{figureFormat}{if 'saveFigure' is TRUE, what format you want to save it as. Default is 'png'}
}
\value{
list of length 6: plotTitle (title of plot); coveragePlot (ggplot of coverage tracks); peakPlot (ggplot of peak tracks); loopPlot (ggplot of loop tracks); genePlot (ggplot of gene/transcript models); figure (patchwork object with all tracks);
}
\description{
plot gene models, coverage tracks, peaks, and/or loops across region of interest
}
