% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{plot_loops}
\alias{plot_loops}
\title{plot loop tracks}
\usage{
plot_loops(
  genomicLoc,
  loopFiles,
  loopTrackNames = NULL,
  loopTrackColors = "black",
  lineSize = 0.8,
  alpha = 0.8,
  rasterizePlot = FALSE,
  specialLoops = NULL,
  specialLoopColors = NULL,
  minScore = NULL,
  loop_orientation = "above",
  fontSize = 9
)
}
\arguments{
\item{genomicLoc}{character vector of the form c(chromosome, start coordinate, end coordinate) (ex: c("chrX", 24040226, 24232664))}

\item{loopFiles}{character vector containing a list of file paths to any contact/loop files (.bedpe format)}

\item{loopTrackNames}{character vector containing the label of each loop track. Should be the same length as 'contactFiles' and in the same order. By default it will label them as "Loops_1", "Loops_2", etc}

\item{loopTrackColors}{character vector containing color of each loop track. If only one color is provided, all tracks will be that color. If different tracks must be different colors, specify the color for each track in order. Default is 'black'}

\item{lineSize}{numeric, width of lines used to draw loops. Default is 0.8}

\item{alpha}{numeric, alpha value (transparency) for loops. Default is 0.8}

\item{rasterizePlot}{boolean, set this to TRUE if you want to rasterize the loop plot. This can be useful as loop files can be rather large, and that can mess up the plot when you save it as a vector file. Default is FALSE}

\item{specialLoops}{character vector listing the names of specific loops you want to put in a different color}

\item{specialLoopColors}{character vector containing the color(s) you want to give your special loops. requires 'specialLoops' to be specified}

\item{minScore}{numeric; any loops with scores lower than this will be thinner and more transparent, making the loops that pass this threshold more visible}

\item{loop_orientation}{either "above" or "below". "above" will draw loops above a horizontal axis; "below" will draw loops below a horizontal axis. Default is "above'}

\item{fontSize}{a numeric for desired font size. Default is 9.}
}
\value{
ggplot of loop tracks
}
\description{
plot loop tracks
}
