% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{plot_peaks}
\alias{plot_peaks}
\title{plot peak tracks}
\usage{
plot_peaks(
  genomicLoc,
  peakFiles,
  peakTrackNames = NULL,
  labelAllPeaks = FALSE,
  peakTrackColors = "black",
  specialPeaks = NULL,
  labelSpecialPeaks = FALSE,
  specialPeakColors = NULL,
  labelStrand = FALSE,
  strandColors = NULL,
  fontSize = 9
)
}
\arguments{
\item{genomicLoc}{character vector of the form c(chromosome, start coordinate, end coordinate) (ex: c("chrX", 24040226, 24232664))}

\item{peakFiles}{character vector containing a list of file paths to any peak files (.bed format)}

\item{peakTrackNames}{character vector containing the label of each peak track. Should be the same length as 'peakFiles' and in the same order. By default it will label them as "Peaks_1", "Peaks_2", etc}

\item{labelAllPeaks}{boolean, set to TRUE if each individual peak should be named. assumes names are provided in the fourth column of a given .bed file}

\item{peakTrackColors}{character vector containing color of each peak track. If only one color is provided, all tracks will be that color. If different tracks must be different colors, specify the color for each track in order. Default is "black"}

\item{specialPeaks}{character vector listing the names of specific peaks you want to either label or put in a different color}

\item{labelSpecialPeaks}{boolean, set to TRUE if each special peak should be named. requires 'specialPeaks' to be specified}

\item{specialPeakColors}{character vector containing the color(s) you want to give your special peaks. requires 'specialPeaks' to be specified}

\item{labelStrand}{boolean, set to TRUE if the strand of each peak should be indicated. assumes strand information is in the sixth column of a given .bed file}

\item{strandColors}{character vector of length 2. First is the color for the + strand, the second is the color for the - strand. overrides 'peakTrackColors' but does NOT override 'specialPeakColors'}

\item{fontSize}{a numeric for desired font size. Default is 9.}
}
\value{
ggplot of peak tracks
}
\description{
plot peak tracks
}
