% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{read_bedgraph_region}
\alias{read_bedgraph_region}
\title{use import or tabular import for bedgraph files}
\usage{
read_bedgraph_region(file, chrom, start_region, end_region)
}
\arguments{
\item{file}{bedgraph file to read}

\item{chrom}{chromosome for the data you want to import}

\item{start_region}{start coordinate}

\item{end_region}{end coordinate}
}
\value{
dataframe of data from bedgraph
}
\description{
use import or tabular import for bedgraph files
}
