% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDJ.R
\name{trackDJ}
\alias{trackDJ}
\title{rearrange outputs from plot_genomic_tracks in the desired order}
\usage{
trackDJ(
  plotList,
  plotOrder,
  saveFigure = FALSE,
  figureFormat = "png",
  figureName = "trackDJ_figure"
)
}
\arguments{
\item{plotList}{a list of outputs from plot_genomic_tracks}

\item{plotOrder}{character vector specifying the order of each track, using the names you provided to plot_genomic_tracks}

\item{saveFigure}{boolean, whether you want to save the final figure to a file. Default is FALSE.}

\item{figureFormat}{if 'saveFigure' is TRUE, what format you want to save it as. Default is 'png'}

\item{figureName}{if 'saveFigure' is TRUE, what you want the filename to be. Default is "trackDJ_figure"}
}
\value{
list of length two: figure (patchwork object with all tracks); singlePlots (list of ggplot objects of each individual track)
}
\description{
rearrange outputs from plot_genomic_tracks in the desired order
}
